source/detector_backend/requirements.txt
numpy==1.15.4
opencv-python==3.4.4.19
Pillow==5.3.0
pyblake2==1.1.2
